# BA/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: lttd0204@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/products/{productId}:
    get:
      tags:
        - products
      summary: Get details product by ID
      description: Get details product by id of products
      operationId: getProductsById
      parameters:
        - name: productId
          in: path
          description: ID of product that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - products
      summary: Update an existing produc
      description: Update an existing product by Id
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: product id of product updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
          application/xml:
            schema:
              $ref: "#/components/schemas/Product"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    delete:
      tags:
        - products
      summary: Delete product by ID
      description: Try delete product by id of products
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: ID of the product that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/products/like:
    post:
      tags:
        - products
      summary: Like product by idProduct
      description: Like product by idProduct
      operationId: likeProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
                  example: admin@gmail.com
              required:
                - productId
          application/xml:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required:
                - productId
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required:
                - productId
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/products/unlike:
    post:
      tags:
        - products
      summary: Unike product by idProduct
      description: Unlike product by idProduct
      operationId: unlikeProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
                  example: admin@gmail.com
              required:
                - productId
          application/xml:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required:
                - productId
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required:
                - productId
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        image:
          type: string
        location:
          type: string
          format: uuid
        price:
          type: number
        countInStock:
          type: number
        description:
          type: string
        discount:
          type: number
        sold:
          type: number
        type:
          type: string
          format: uuid
        discountStartDate:
          type: string
          format: date-time
        discountEndDate:
          type: string
          format: date-time
        status:
          type: number
          default: 0
          enum: [0, 1]
      required:
        - name
        - slug
        - image
        - type
        - price
        - countInStock
        - location
      xml:
        name: Product
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT